admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "web", from = "src" }]

[tool.poetry.dependencies]
python = "3.12.3"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
flask-smorest = "^0.44.0"
flask-session = "^0.8.0"
flask-bcrypt = "^1.0.1"
minio = "^7.2.9"
ulid-py = "^1.1.0"
bleach = "^6.1.0"
authlib = "^1.3.2"
requests = "^2.32.3"
autopep8 = "^2.3.1"
python-dateutil = "^2.9.0.post0"
flask-cors = "^5.0.0"




[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
